⚡️ Set DEFAULT_D to 6, based on benchmarks.
This is based on results for `benchmarks/n_push_n_pop_n.yml`.  And `d=8`
is nearly identical.

The results for `benchmarks/n_push_pop.yml` can be thrown off a little
bit, based on what the specific N value does for the size of the last
set of children (and how that interacts with e.g.  compiler unrolled
loops).  `push_n_pop_n` isn't as affected by that because it already
hits every possible `n % d` size.  To compensate for this, I'd need to
add some "jitter" into that test so it measures every possible size for
the last children.