♻️ ⚡️ short-circuit for the default D
Even better would be to short circuit for the *current* D, and use a fn
pointer for sift-down (or during object construction).  Using a fn
pointer for min_index was a significant slowdown, but ruby's
object/method polymorphism is essentially a complicated fn pointer, so
putting it at that level should have no downside (or near that level
would have negligible downside).